requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1

matplotlib==3.5.3
setuptools

hydra-core
hydra_colorlog

seaborn
torch>=2.0.1
torchvision
lightning
pytorch-lightning
mlflow
numpy
scikit-learn
pyro-ppl

omegaconf
tqdm
ray[tune]==1.13.0

git+https://github.com/fadel/pytorch_ema